Only deletes temp credentials file

+++ src/python/aqueduct_executor/operators/connectors/data/gcs.py
@@ -14,6 +14,7 @@ _CREDENTIALS_ENV_VAR = "GCS_CREDENTIALS"
 class GCSConnector(connector.DataConnector):
     _client: Any  # GCS client
     _config: GCSConfig
+    _temp_credentials_path: str = ""
 
     def __init__(self, config: GCSConfig):
         if _CREDENTIALS_ENV_VAR in os.environ in os.environ:
@@ -21,15 +22,18 @@ class GCSConnector(connector.DataConnector):
             temp_path = os.path.join(os.getcwd(), str(uuid.uuid4()))
             with open(temp_path, "w") as f:
                 f.write(os.environ[_CREDENTIALS_ENV_VAR])
+            
             config.credentials_path = temp_path
+            self._temp_credentials_path = temp_path
 
         self._client = storage.Client.from_service_account_json(config.credentials_path)
         self._config = config
 
     def __del__(self):
-        # Try to clean up temp credentials file
-        os.remove(self._config.credentials_path)
-
+        if self._temp_credentials_path:
+            # Try to clean up temp credentials file
+            os.remove(self._config.credentials_path)
+            
     def authenticate(self) -> None:
         self._client.list_buckets()
